requirements.txt
flask==3.1.*
requests==2.32.*
beautifulsoup4==4.13.*
websocket-client==1.8.*
